Added functionality to output form M2 script
The M2-code generated in eqs2M2, now also generates a text file with
dimension, nr of solutions and leading terms of the grabber basis.
Added the script runM2 that takes equations, generates a
Macauley-script, invokes this script and returns the dimension, nr of
solutions and leading terms of the polynomials in the grabber basis.